Fix cloudwatch rule target ARN
AWS is inconsistent about whether a cloudwatch ARN has :* appended to the end, and we need to make sure it doesn't in the rule target. Was creating a rule that didn't succesfully send events anywhere before.

This weird thing with cloudwatch ARNs seems to be a general inconistency in AWS API/SDK stack, compare to https://github.com/aws/aws-cdk/issues/8137