MediaConvert adapter create has output_group_destination_settings config
You can supply a hash that will be passed to MediaConvert destination_settings in the relevant output group configuration.

I needed this for my app for a use case where I need to tell MediaConvert to use canned PUBLIC_READ acl when creating files. I could have created a feature just for like `canned_acl`, but it was simpler to implement and seemed convenient to let the caller pass in any `destination_settings`; the only other one that exists is basically S3 encryption, but, hey, maybe somebody wants that.

In addition to writing a mocked test, I have tested this with my live app, it works.