Refactor check_s3_bucket method
Co-authored-by: Chris Colvard <chris.colvard@gmail.com>

If a file is already successfully in an s3 bucket and is not being moved
to a new bucket we do not need to do anything with the filename. If the
file is being moved to a new bucket, we should sanitize it to the
stricter set of allowed characters.